Added source dir argument
Now if a different source directory is used for html files, it can be
sourced, otherwise `./html/` will be used. Also, if a file cannot be
found, it will send a 404 with a special "not found" html. This had to
be put into the source code in because the reading of a file could
always fail.