TODO: What to do if the logger fails?
Added a `TODO` message for now, it is not really clear to me what should
be done when the logger of the program fails. On the one hand, it is not
critical to the program, on the other hand, it might cause the program
to silently start failing with no indication. Maybe if it fails just
`eprint`? Will think about this and implement a change when I have come
to a decision.