Version bump
I have bumped the version up in `Cargo.toml`. The server has full basic
functionality. It can send files, parse reqests, and log everything.
There is also a way to edit some of the defaults of the server such as
which address and port to use, logger options, and where to look for
files to send as a response. There is still more to be done (such as
gracefull shutdown, checking permissions of the directory used, and
others) but after re-organising some things it feels ready for a version
bump.