Moved the listener loop to its own function
The `listener.incoming()` loop has been moved to its own function. The
loop has also been put in a thread to prepare it for later mangement.
The loop itself has also been modified, now there is an atomic boolean
variable `stop` which is checked to see if the thread should stop. There
is, however, a potential "problem" (also pointed out in a comment). The
`incoming()` function is blocking, meaning that the `stop` variable is
checked only when a request comes in. This means that there might be a
situation to force stop the thread if it hasn't finished within some
period.