Decided to remove all white-space trimming from config file
This is done in order to be able to process strings properly without
needing to escape them (for example with `"`). It means that the format
is now required to be `key=value` without any spaces, as spaces are
treated as part of the key and value, respectivelly, it does make
parsing simpler while being able to have strings end and start with
spaces.